+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Additions
 
+- UIComponent.forwardWorkspaceActions(target) will override the current receiver of button states and action callbacks
 - replaceWith(component: KClass<T>) accepts `sizeToScene` and `centerOnScreen`
 - titledpane builder that accepts the title as ObservableValue<String>
 - TaskStatus.completed and FXTask.completedProperty can be used to listen to changes in completion state of a task